refactor: Enum의 Converter를 내부 클래스로 이동
