+++ src/main/java/kr/composite/api/quiz/domain/QuizStatus.java
@@ -1,10 +1,10 @@
 package kr.composite.api.quiz.domain;
 
-import lombok.Getter;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
 
 import java.util.Arrays;
 
-@Getter
 public enum QuizStatus {
 
     NOT_STARTED("시작 전"),
@@ -17,10 +17,28 @@ public enum QuizStatus {
         this.description = description;
     }
 
-    public static QuizStatus from(String description) {
-        return Arrays.stream(values())
-                .filter(value -> value.description.equals(description))
-                .findFirst()
-                .orElseThrow(IllegalArgumentException::new);
+    @Converter
+    public class QuizStatusConverter implements AttributeConverter<QuizStatus, String> {
+
+        @Override
+        public String convertToDatabaseColumn(QuizStatus quizStatus) {
+            if (quizStatus == null) {
+                return null;
+            }
+
+            return quizStatus.description;
+        }
+
+        @Override
+        public QuizStatus convertToEntityAttribute(String dbData) {
+            if (dbData == null) {
+                return null;
+            }
+
+            return Arrays.stream(values())
+                    .filter(value -> value.description.equals(description))
+                    .findFirst()
+                    .orElseThrow(IllegalArgumentException::new);
+        }
     }
 }
